fix(money): re-land the Decimal migration — 210 float-money violations to 0 [AITBC-152]
Squashes PRs #161-#165, which Gitea marked merged but whose content never
reached main. Verified by content rather than SHA: main still had
no_float_money_baseline.json at "total": 210, no cli/aitbc_cli/utils/money.py,
none of the four Alembic revisions, and a v0.23 release log ending at V23-33.
The commits survived only as dangling objects; they are tagged recovered/* .

Rebased onto current main (through PR #179) and applied with no conflicts.
V23-34 through V23-40 in docs/releases/v0.23/release.log.

  coordinator-api  168 -> 0   c7d1f4a9e230 converts 102 columns across 45 tables
  blockchain-node   24 -> 0   d4e8b91c0a37 converts two GPU resource columns
  cli               11 -> 0
  the rest          13 -> 0   f2b6c04a91d8, a3e7c15b8d94
  parameters       131 -> 0

Six widenings of scripts/lint/no_float_money.py, each of which found real
defects in code the previous pass had reported clean: annotations; `per` and
`target` in DERIVED_TOKENS hiding price_per_hour and target_amount; camelCase;
ast.IfExp; ast.Attribute annotation targets hiding self.earnings and
self.total_spent; and parameter annotations. "Zero violations" only ever means
zero of what the checker can currently see, so the guard now says so in its
docstring and the baseline records that it started at 210.

The largest single defect was in ait_to_seconds, which turns a user's --amount
into the integer compute-seconds the chain settles:

    int(0.5025 * 3600) == 1808        0.5025 * 3600 == 1808.9999999999998
    int(Decimal("0.5025") * 3600) == 1809

1402 of the million four-decimal inputs between 0.0001 and 100.0000 lost a
compute-second that way, always the sender's. tests/unit/test_ait_unit_conversion.py
pins both the fix and the old behaviour.

@click.option(type=float) made CLI annotations unenforceable, since Click
converts before the handler runs, so cli/aitbc_cli/utils/money.py adds a
DecimalParamType used by every money option. Local wallet, multisig and
exchange JSON now store money as decimal strings, read back through
wallet_amount(), which also accepts the numbers older builds wrote.

Money that crosses a signature or hash boundary deliberately stays float, each
site carrying a `# not-money:` marker with its reason -- Decimal is not
JSON-serialisable and "0.5" != 0.5 changes the hash, so converting those is a
hard fork. Same for quantities whose names carry a money token but which are
rates, ranking weights, timeouts, RL returns or numpy statistics.

ruff, mypy aitbc/, the 12-app mypy ratchet and the guard all pass. 2672 passed
across tests/unit+cli+services+security; blockchain-node 699, trading 84, miner
60, agent-coordinator 76, pool-hub 41, agent-sdk 25, gpu 17, edge 11, crypto 2.
coordinator-api's 48 failures are byte-identical to plain origin/main, verified
in a detached worktree.

+++ apps/coordinator-api/src/coordinator_api/contexts/agent_coordination/services/task_decomposition.py
@@ -5,6 +5,7 @@ Implements intelligent task splitting and sub-task management
 
 from dataclasses import dataclass, field
 from datetime import UTC, datetime
+from decimal import Decimal
 from enum import StrEnum
 
 from aitbc.aitbc_logging import get_logger
@@ -69,7 +70,7 @@ class TaskRequirement:
     data_size: int
     priority: int
     deadline: datetime | None = None
-    max_cost: float | None = None
+    max_cost: Decimal | None = None
 
 
 @dataclass
@@ -103,7 +104,7 @@ class TaskDecomposition:
     dependency_graph: dict[str, list[str]]
     execution_plan: list[list[str]]
     estimated_total_duration: float
-    estimated_total_cost: float
+    estimated_total_cost: Decimal
     confidence_score: float
     decomposition_strategy: str
     created_at: datetime = field(default_factory=lambda: datetime.now(UTC))
